Add one_file_system preference flag (for *nix)
Not currently supported on Windows (turns out this is a PITA which
will require so reworking to avoid slowing down things terribly -
ie like most things in Windows you basically have to open the file
and opening files in general there is already slow, and right now it
will be difficult to reuse any handles with the current architecture.

On *nix systems it simply uses the st_dev field of the stat buffer.